+++ src/PlayedCardButton.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import noEvent from "./sound-effects/noEvent.mp3";
 import yesEvent from "./sound-effects/yesEvent.mp3";
 import { pulsar } from "ldrs";
@@ -7,16 +7,30 @@ pulsar.register();
 
 export default function PlayedCardButton({
   setEventTriggered,
+  eventType,
   setEventType,
   isSoundEffectMuted,
   numberOfPlayers,
+  playerNames,
+  setEnvironment,
   environment,
+  setHeader,
+  setPlayersInvolved,
+  setInstructions,
+  setPlayContinue,
 }) {
   const [isLoading, setIsLoading] = useState(false);
   //state to keep of how many times an event has not happened in a row
   const [countDown, setCountDown] = useState(4);
   //state to keep of event happening so it does not happen twice in a row
   const [eventTriggeredCooldown, setEventTriggeredCooldown] = useState(true);
+
+  useEffect(() => {
+    if (eventType) {
+      generateEventDetails();
+      setEventTriggered(true);
+    }
+  }, [eventType]);
 
   //generate a number between 0 to 100
   function randomNumberGenerator() {
@@ -74,8 +88,309 @@ export default function PlayedCardButton({
       setEventType(selectedEvent);
     } else if (eventOrEnvironment() === "event") {
       selectedEvent = events[Math.floor(Math.random() * events.length)];
+      console.log(selectedEvent);
       setEventType(selectedEvent);
     }
+  }
+
+  function generateEventDetails() {
+    console.log(`eventtype is ${eventType}`);
+    let text = "";
+
+    function generateTwoUniqueRandomNumbers(array) {
+      // Get the size of the array
+      const arraySize = array.length;
+
+      // Generate the first random number
+      const randomNumber1 = Math.floor(Math.random() * arraySize);
+
+      // Generate the second random number, ensuring it's different from the first one
+      let randomNumber2;
+      do {
+        randomNumber2 = Math.floor(Math.random() * arraySize);
+      } while (randomNumber2 === randomNumber1);
+
+      // Return the two random numbers
+      return [randomNumber1, randomNumber2];
+    }
+    const cardTypes = [
+      "green",
+      "red",
+      "blue",
+      "yellow",
+      "special",
+      "odd",
+      "even",
+    ];
+
+    const colourTypes = ["green", "red", "blue", "yellow"];
+
+    const [randomNumber1] = generateTwoUniqueRandomNumbers(playerNames);
+
+    const randomCardType =
+      cardTypes[Math.floor(Math.random() * cardTypes.length)];
+
+    const randomColourType =
+      colourTypes[Math.floor(Math.random() * colourTypes.length)];
+
+    const randomCardNumber = Math.ceil(Math.random() * 3);
+
+    let header = (
+      <h1>
+        <span className="underline">{eventType}!</span>
+      </h1>
+    );
+
+    setHeader(header);
+
+    let playersInvolved;
+    let instructions;
+    if (eventType.toLowerCase() == "luck of the draw") {
+      playersInvolved = (
+        <>
+          <p>
+            The player involved is{" "}
+            <span className="font-bold underline">
+              {playerNames[Math.floor(Math.random() * playerNames.length)]}
+            </span>
+          </p>
+        </>
+      );
+      instructions = (
+        <>
+          <p>
+            The player must draw cards until they get{" "}
+            {/^[aeiouAEIOU].*/.test(randomCardType) ? "an" : "a"}{" "}
+            <span className="font-bold underline">{randomCardType}</span> card!
+          </p>
+        </>
+      );
+    } else if (eventType.toLowerCase() == "swapparoo") {
+      const [randomNumber1, randomNumber2] =
+        generateTwoUniqueRandomNumbers(playerNames);
+      playersInvolved = (
+        <>
+          <p>
+            The players involved are{" "}
+            <span className="font-bold underline">
+              {playerNames[randomNumber1]}
+            </span>{" "}
+            and{" "}
+            <span className="font-bold underline">
+              {playerNames[randomNumber2]}
+            </span>
+          </p>
+        </>
+      );
+      instructions = (
+        <p>The players must swap all their cards with each other!</p>
+      );
+    } else if (eventType.toLowerCase() == "trade & bargain") {
+      const [randomNumber1, randomNumber2] =
+        generateTwoUniqueRandomNumbers(playerNames);
+      playersInvolved = (
+        <>
+          <p>
+            The players involved are{" "}
+            <span className="font-bold underline">
+              {playerNames[randomNumber1]}
+            </span>{" "}
+            and{" "}
+            <span className="font-bold underline">
+              {playerNames[randomNumber2]}
+            </span>
+          </p>
+        </>
+      );
+      instructions = (
+        <p>
+          The players must swap{" "}
+          <span className="underline font-bold">{randomCardNumber}</span>{" "}
+          {randomCardNumber > 1 ? "cards " : "card "} with each other!
+        </p>
+      );
+    } else if (eventType.toLowerCase() == "wormhole") {
+      playersInvolved = (
+        <>
+          <p>
+            The player involved is{" "}
+            <span className="font-bold underline">
+              {playerNames[Math.floor(Math.random() * playerNames.length)]}
+            </span>
+          </p>
+        </>
+      );
+      instructions = (
+        <p>
+          The player must remove all{" "}
+          <span className="font-bold underline">
+            {cardTypes[Math.floor(Math.random() * cardTypes.length)]}
+          </span>{" "}
+          cards from their hand!
+        </p>
+      );
+    } else if (eventType.toLowerCase() == "flip & fill") {
+      playersInvolved = (
+        <>
+          <p>
+            The player involved is{" "}
+            <span className="font-bold underline">
+              {playerNames[Math.floor(Math.random() * playerNames.length)]}
+            </span>
+          </p>
+        </>
+      );
+      instructions = (
+        <p>
+          The player must flip the next card in the draw pile and draw that
+          number of cards. Special cards means zero.
+        </p>
+      );
+    }
+    // else if (eventType.toLowerCase() == "pick & pray") {
+    //   const [randomNumber1, randomNumber2] =
+    //     generateTwoUniqueRandomNumbers(playerNames);
+    //   let randomNumber3 = Math.floor(Math.random() * playerNames.length);
+    //   if (randomNumber3 === randomNumber1 || randomNumber3 === randomNumber2) {
+    //     playersInvolved = (
+    //       <>
+    //         <p>
+    //           The players involved are{" "}
+    //           <span className="font-bold underline">
+    //             {playerNames[randomNumber1]}
+    //           </span>{" "}
+    //           and{" "}
+    //           <span className="font-bold underline">
+    //             {playerNames[randomNumber2]}
+    //           </span>{" "}
+    //         </p>
+    //       </>
+    //     );
+    //   } else {
+    //     playersInvolved = (
+    //       <>
+    //         <p>
+    //           The players involved are{" "}
+    //           <span className="font-bold underline">
+    //             {playerNames[randomNumber1]}
+    //           </span>{" "}
+    //           ,{" "}
+    //           <span className="font-bold underline">
+    //             {playerNames[randomNumber2]}
+    //           </span>{" "}
+    //           and{" "}
+    //           <span className="font-bold underline">
+    //             {playerNames[randomNumber3]}
+    //           </span>
+    //         </p>
+    //       </>
+    //     );
+    //   }
+
+    //   instructions = (
+    //     <p>
+    //       <span className="font-bold underline">
+    //         {playerNames[randomNumber1]}
+    //       </span>{" "}
+    //       must take a card from{" "}
+    //       <span className="font-bold underline">
+    //         {playerNames[randomNumber2]}
+    //       </span>
+    //       .{" "}
+    //       <span className="font-bold underline">
+    //         {playerNames[randomNumber3]}
+    //       </span>{" "}
+    //       must draw cards equal to the taken card. Special cards means zero.
+    //     </p>
+    //   );
+    // }
+    else if (eventType.toLowerCase() == "whopping wormhole") {
+      playersInvolved = (
+        <>
+          <p>Everyone is affected!</p>
+        </>
+      );
+      instructions = (
+        <p>
+          Everyone must remove all{" "}
+          <span className="font-bold underline">
+            {cardTypes[Math.floor(Math.random() * cardTypes.length)]}
+          </span>{" "}
+          cards from their hand!
+        </p>
+      );
+    } else if (eventType.toLowerCase() == "shade shuffle") {
+      playersInvolved = (
+        <>
+          <p>Everyone is affected!</p>
+        </>
+      );
+      instructions = (
+        <p>
+          The colour in play has now been changed to{" "}
+          <span className="font-bold underline">{randomColourType}</span>
+        </p>
+      );
+    } else if (eventType.toLowerCase() == "new neighbours") {
+      const [randomNumber1, randomNumber2] =
+        generateTwoUniqueRandomNumbers(playerNames);
+      playersInvolved = (
+        <>
+          <p>
+            The players involved are{" "}
+            <span className="font-bold underline">
+              {playerNames[randomNumber1]}
+            </span>{" "}
+            and{" "}
+            <span className="font-bold underline">
+              {playerNames[randomNumber2]}
+            </span>
+          </p>
+        </>
+      );
+      instructions = (
+        <p>
+          The players must{" "}
+          <span className="font-bold underline">swap seats</span> with each
+          other!
+        </p>
+      );
+    } else if (eventType.toLowerCase() == "silent library") {
+      setEnvironment("silent library");
+      playersInvolved = (
+        <>
+          <p>New Environment!</p>
+        </>
+      );
+      instructions = <p>The game must now continue in complete silence.</p>;
+    } else if (eventType.toLowerCase() == "mafia manor") {
+      setEnvironment("mafia manor");
+      playersInvolved = (
+        <>
+          <p>New Environment!</p>
+        </>
+      );
+      instructions = <p>All pick up card effects are doubled.</p>;
+    } else if (eventType.toLowerCase() == "cosy campsite") {
+      setEnvironment("cosy campsite");
+      playersInvolved = (
+        <>
+          <p>New Environment!</p>
+        </>
+      );
+      instructions = <p>All pick up card effects are halved.</p>;
+    }
+    let playContinue = (
+      <p>
+        Play will continue from{" "}
+        <span className="font-bold underline">
+          {playerNames[randomNumber1]}
+        </span>
+      </p>
+    );
+    setPlayersInvolved(playersInvolved);
+    setInstructions(instructions);
+    setPlayContinue(playContinue);
   }
 
   //count down to keep track of how many times an event has not happened in a row
@@ -87,7 +402,6 @@ export default function PlayedCardButton({
 
     if ((randomNumber <= 25 && !eventTriggeredCooldown) || countDown == 0) {
       playYesEvent();
-      setEventTriggered(true);
       randomEventSelect();
     } else {
       setEventTriggeredCooldown(false);
